CoroutineScope をコンストラクタで渡すと singleton にできないので、開発用の処理のために、CoroutineScope を渡すのをやめる
