+++ Sample/MultipleModuleSample/app/src/main/java/com/github/mag0716/multiplemodulesample/fragment/ListFragment.kt
@@ -23,7 +23,7 @@ class ListFragment : Fragment(), CoroutineScope {
 
     private lateinit var job: Job
 
-    private val getDataListUseCase = provideGetDataListUseCase(this)
+    private val getDataListUseCase = provideGetDataListUseCase()
 
     private lateinit var adapter: Adapter
