+++ Sample/MultipleModuleSample/datasource/src/main/java/com/github/mag0716/datasource/DataSourceModule.kt
@@ -1,9 +1,12 @@
 package com.github.mag0716.datasource
 
-import kotlinx.coroutines.experimental.CoroutineScope
-
 object DataSourceModule {
-    fun provide(coroutineScope: CoroutineScope): IDataRepository {
-        return DataRepository(coroutineScope)
+
+    private val dataRepository: IDataRepository by lazy {
+        DataRepository()
+    }
+
+    fun provide(): IDataRepository {
+        return dataRepository
     }
 }
