+++ Sample/MultipleModuleSample/usecase/src/main/java/com/github/mag0716/usercase/UseCaseModule.kt
@@ -1,12 +1,11 @@
 package com.github.mag0716.usercase
 
 import com.github.mag0716.datasource.DataSourceModule
-import kotlinx.coroutines.experimental.CoroutineScope
 
-fun provideGetDataListUseCase(coroutineScope: CoroutineScope): IGetDataListUseCase {
-    return GetDataListUseCase(DataSourceModule.provide(coroutineScope))
+fun provideGetDataListUseCase(): IGetDataListUseCase {
+    return GetDataListUseCase(DataSourceModule.provide())
 }
 
-fun provideGetDataDetailUseCase(coroutineScope: CoroutineScope): IGetDataDetailUseCase {
-    return GetDataDetailUseCase(DataSourceModule.provide(coroutineScope))
+fun provideGetDataDetailUseCase(): IGetDataDetailUseCase {
+    return GetDataDetailUseCase(DataSourceModule.provide())
 }
